Project skeleton

+++ symlink-lib.sh
@@ -23,8 +23,10 @@ echo "- - -"
 
 if [ -h "$dest/errors" ]; then
   for entry in $LIBRARIES_DIR/$name*; do
-    echo "- removing symlink: $entry"
-    rm $entry
+    if [ -e "$entry" ]; then
+      echo "- removing symlink: $entry"
+      rm $entry
+    fi
   done
 fi
